chore(): refactor app module to be located at src root
